Add agent class to handle login/logout
- Make command methods return command only, without executing.
- Add class-level Command.run
- Use capture2e for run
- Add default configuration parameter for executable
- Prevent login from running when already logged-in
- Prevent logout from running when already logged-out
