+++ lib/lastpass/cli/configuration.rb
@@ -1,7 +1,11 @@
 module Lastpass
   module CLI
     class Configuration
-      attr_accessor :username, :password
+      attr_accessor :username, :password, :executable
+
+      def initialize
+        @executable = 'lpass'
+      end
     end
   end
 end
